기능개발 : 워크스페이스 화면 컴포넌트 분리 및 생성, 수정, 삭제 기능 개발
