+++ react-app/src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import reportWebVitals from './reportWebVitals';
 import App from './App';
-import { UserProvider } from './component/hook/UserProvider';
+import { UserProvider } from './components/hooks/UserProvider';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
